requirements.txt
streamlit
python-dotenv
groq
faiss-cpu
sentence-transformers